refactor: Restructure project layout and extract mock data
- Moved GraphView component to a dedicated graph directory
- Created MainLayout component for consistent page structure
- Extracted mock graph data to a separate constants file
- Updated import paths and component references
- Relocated global CSS to styles directory
- Improved project organization and separation of concerns
