+++ frontend/src/App.test.jsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import App from './App';
 
 // Mock the GraphView component to avoid D3 initialization
-jest.mock('./components/GraphView/GraphView', () => {
+jest.mock('./components/graph/GraphView', () => {
   return function MockGraphView() {
     return <div data-testid="mock-graph">Graph View</div>;
   };
