+++ frontend/src/components/graph/GraphView.jsx
