+++ frontend/src/components/layout/MainLayout.module.css
@@ -1,0 +1,28 @@
+.mainLayout {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  font-size: 3.2em;
+  line-height: 1.1;
+  margin-bottom: 2rem;
+  background: linear-gradient(120deg, #646cff, #ff7f50);
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  background-clip: text;
+}
+
+.content {
+  width: 100%;
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
