+++ frontend/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import './index.css';
+import './styles/index.css';
 import App from './App';
 
 createRoot(document.getElementById('root')).render(
